Update afros-Harmony-Add-Ons.user.js
Another pretty big update!
Everything works pretty much the same, but I've re written a lot of the code, especially the code to copy the list of countries.

Most noticeable changes are that you can now copy the list of available countries, and the catalog number as well.

changelog:
- bumped to version 1.2
- added custom css classes instead of relying on inline styling
- replaced all instances of `var` with either `let` or `const` 
- rewrote the qobuz code, now uses an object instead of a map to get the relevant region. also reduced the entire `if ... else` logic to pretty much a single line using `.map` and `.find` 
- replaced the constructed search URLs with template literals
- added some comments that should make future maintenance easier
- realized I don't need to add two different event listeners to elements, so moved the notification triggers and the copying to clipboard to the same event listener
- moved all of the `.mouseover / .mouseout` functions that were just changing text color to a simple `:hover` css
- Rewrote the entire `copyCountries()` function. Now instead of doing everything in a single function, it's split into a few different functions handling different tasks. This happened because I realized what a mess this would be to enable copying the "Availability" list.
A lot of the code was cut down because I also realized Harmony assigns the class `.region-list` to the lists, so this cuts down on having to find them by iterating through the table to find the one cell with the "unavailability" text and so on.
Also found out I can get the lists by checking the `abbr` tags on the flags when the lists are collapsed. It's much easier to get the lists on page load and not have to worry about the lists being expanded or not. 
- added the ability to copy the catalog number in a single click as well. The blue background and dotted underline should make it obvious that it's copyable
- overall reduced the code by about a hundred lines(!!) or about 5k characters

In the next update I'll probably change the permalink href `javascript:void(0)` to something else, maybe just a clickable button instead of a link. See https://stackoverflow.com/a/1293130