Add header text & digital store support
Pretty big update :D
Main things: I've added a "title" header to the sublinks window, and links to digital stores (spotify, apple music, deezer, tidal, and beatport) get a sublinks window with options to lookup with Harmony or a-tisket (where applicable)
Discogs links also get a sublinks window with the option of checking the discogs edit history (need a discogs account to see it though)

I've refrained from using the MB API but I thought about it when I had to deal with MB links that don't have the "title" of the link as its text, like it happens when using MB links in edit notes. The function `getHeaderText` deals with this by using the generic entity name instead of the actual name of the entity, for example, it will show "Artist" instead of "Kill the Noise" when opening the sublinks window over a link to https://musicbrainz.org/artist/a9930a14-8fbb-435a-ad25-f4b6cae71523 in an edit note.
I think if I implement API calls it would make the script way slower for just a simple aesthetic reason.

Next on my to-do list is figuring out how to keep the sublinks window inside the page window when activating near the edges. That'll come in the next update (probably)