Handle SPA navigation
Pretty much entirely new code, it took me a while to figure out how to deal with beatsource's navigation shenanigans
The script will now automatically get the right data when navigating to another release. It no longer needs to make every <a> element reload the page (what the previous linkMe() function did)

Other than the navigation fixes, not much has changed. The harmony link now looks like a link instead of plain text. I also made it so that if the catalog number is exactly the same as the barcode, it'll just skip showing it at all, since it's likely not a real catalog number